server/requirements.txt
djangorestframework==3.10.3
django==2.2.7
djangorestframework-jwt==1.11.0
django-modeltranslation==0.14
six==1.13.0
pytest-django==3.7.0
django-cors-headers==3.2.0